Information listings now fetch and update the state.
