+++ src/components/ListingContainer/listingContainer.js
@@ -1,12 +1,8 @@
 import React, { Component } from 'react';
-import { ListingCard } from '/Users/trondmakonese/mod_3/V-RAD/vrad-project/src/components/ListingCard/listingCard.js'
+// import { ListingCard } from './components/ListingCard/listingCard.js'
 
 export const ListingContainer = (props) => {
-    const allListings = props.listingByArea.map(area => {
-        return area.listings.map(listing => {
-            return props.getListings(listing)
-        })
-    })
+    console.log(props)
 
     return (
     <section>
@@ -15,21 +11,3 @@ export const ListingContainer = (props) => {
     )
 }
 
-
-// export default class ListingContainer extends Component {
-//     constructor(props){
-//         super(props);
-//         this.state = {
-//             listings: []
-//         }
-//     }
-
-//     render(props) {
-//         console.log(this.props)
-//         return (
-//         <section>
-//             {allListings}
-//         </section>
-//         )
-//     }
-// }
